Tests: Pass PID of child process to parent
This fixes the "ReadProcessMemory" failure in the MiniDumpWriteDump() call.

The root issue is that Cargo is being used to run the test executable, and
so `child.id()` returns Cargo's PID and not the PID of the test.exe
executable.

This has the test.exe pass its actual PID back to the parent so it can
be debugged properly.

Note that this doesn't fully fix the test -- There is still another error