Fix failing tests.
The sanitized stacks test was previously not running the `Context::With`
case. Now that the `With`/`Without` tests are programmatically toggled
(to prevent such errors as that!), it is clear that there is something
weird with the context's stack pointer, which seems to rarely (never?)
lie in mapped memory. For now, this test is disabled.