Write stacks correctly during stack overflows
When encountering a stack overflow we often crash accessing the guard
page. The logic assumed that wherever the stack pointer was so was the
stack, but this lead the writer to dump the guard page in these cases.
This patch changes the logic to inspect the properties of the mapping
that appears to correspond to the stack and - if it looks like a guard
page - look for the actual stack instead.

This change also removes the double limitation we had when retrieving
stacks on Linux: previously the logic would only grab the first 32 KiB
of each stack before checking for user-specified limits. Now only the
user-specified limits are enforced and - if not present - the full
stack is stored in the minidump. This brings the behavior in line with
minidumps generated on Windows by windbg.dll.

This fixes #24